Trait archibald::server::archibaldserver::ServerHandler

pub trait ServerHandler {
    fn handle_request(&mut self, request: &Request<'_>) -> Response;
    fn handle_bad_request(&mut self, e: &ParseError) -> Response;
}

Required Methods

fn handle_request(&mut self, request: &Request<'_>) -> Response

This is the main handler for the server. It will take a request and return a response.

fn handle_bad_request(&mut self, e: &ParseError) -> Response

This is the error handler for the server. It will take a request and return a response.